+++ fixture-monkey-kotlin/src/test/kotlin/com/navercorp/fixturemonkey/kotlin/test/PrimaryConstructorArbitraryIntrospectorTest.kt
@@ -20,10 +20,8 @@ package com.navercorp.fixturemonkey.kotlin.test
 
 import com.navercorp.fixturemonkey.FixtureMonkey
 import com.navercorp.fixturemonkey.LabMonkey
-import com.navercorp.fixturemonkey.api.introspector.BeanArbitraryIntrospector
 import com.navercorp.fixturemonkey.kotlin.KotlinPlugin
 import com.navercorp.fixturemonkey.kotlin.giveMeOne
-import com.navercorp.fixturemonkey.kotlin.spec.JavaObject
 import net.jqwik.api.Property
 import org.assertj.core.api.BDDAssertions.then
 import org.assertj.core.api.BDDAssertions.thenNoException
@@ -80,17 +78,5 @@ class PrimaryConstructorArbitraryIntrospectorTest {
         val actual = sut.giveMeOne<InterfaceClass>()
 
         then(actual).isNull()
-    }
-
-    @Property
-    fun sampleJavaObject() {
-        val sut: LabMonkey = FixtureMonkey.labMonkeyBuilder()
-            .plugin(KotlinPlugin())
-            .objectIntrospector(BeanArbitraryIntrospector.INSTANCE)
-            .build()
-        // when
-        val actual = sut.giveMeOne<JavaObject>()
-
-        then(actual).isNotNull
     }
 }
